+++ js/utils/get-random-positive-integer.js
@@ -1,0 +1,6 @@
+function getRandomPositiveInteger(a, b) {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  return Math.floor(Math.random() * (upper - lower + 1) + lower);
+}
+export { getRandomPositiveInteger };
